Refactor index.ts to parameters.types
