+++ src/app/(root)/orders/page.tsx
@@ -1,7 +1,7 @@
 import Search from '@/components/shared/Search';
 import {getOrdersByEvent} from '@/lib/actions/order.actions';
 import {formatDateTime, formatPrice} from '@/lib/utils';
-import {SearchParamProps} from '@/types';
+import {SearchParamProps} from '@/types/parameters.types';
 import {IOrderItem} from '@/lib/database/models/order.model';
 
 const Orders = async ({searchParams}: SearchParamProps) => {
