+++ src/app/(root)/profile/page.tsx
@@ -3,7 +3,7 @@ import {Button} from '@/components/ui/button';
 import {getEventsByUser} from '@/lib/actions/event.actions';
 import {getOrdersByUser} from '@/lib/actions/order.actions';
 import {IOrder} from '@/lib/database/models/order.model';
-import {SearchParamProps} from '@/types';
+import {SearchParamProps} from '@/types/parameters.types';
 import {auth} from '@clerk/nextjs';
 import Link from 'next/link';
 import React from 'react';
