+++ src/lib/utils.ts
@@ -3,7 +3,7 @@ import {type ClassValue, clsx} from 'clsx';
 import {twMerge} from 'tailwind-merge';
 import qs from 'query-string';
 
-import {UrlQueryParams, RemoveUrlQueryParams} from '@/types';
+import {UrlQueryParams, RemoveUrlQueryParams} from '@/types/parameters.types';
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
