+++ src/types/env.d.ts
@@ -4,11 +4,4 @@ declare namespace NodeJS {
     CLERK_SECRET_KEY: string;
     MONGODB_URI: string;
   }
-
-  // export interface Global {
-  //   mongoose: {
-  //     conn: null | Mongoose,
-  //     promise: Promise<Mongoose>
-  //   };
-  // }
 }
