+++ tsconfig.json
@@ -32,8 +32,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    ".next/types/**/*.ts",
-    "./custom.types.d.ts"
+    ".next/types/**/*.ts"
   ],
   "exclude": [
     "node_modules"
